:sparkles: add localtest&test subcommand for cloud env
