+++ mse_cli/common/helpers.py
@@ -2,19 +2,27 @@
 
 import os
 import shutil
+import subprocess
+import sys
 from pathlib import Path
 from typing import Optional
 
 import pkg_resources
+from docker.errors import NotFound
+from docker.models.containers import Container
 from jinja2 import Template
 
 from mse_cli.cloud.api.types import DefaultAppConfig
+from mse_cli.core.bootstrap import is_ready
+from mse_cli.core.clock_tick import ClockTick
 from mse_cli.core.conf import AppConf, AppConfParsingOption
+from mse_cli.core.test_docker import TestDockerConfig
 from mse_cli.home.model.package import (
     DEFAULT_CODE_DIR,
     DEFAULT_CONFIG_FILENAME,
     DEFAULT_TEST_DIR,
 )
+from mse_cli.log import LOGGER as LOG
 
 
 def scaffold(app_name: str, cloud_config: Optional[DefaultAppConfig] = None) -> Path:
@@ -51,7 +59,7 @@ def scaffold(app_name: str, cloud_config: Optional[DefaultAppConfig] = None) -> 
 
     # Initialize the python code file
     code_dir = project_dir / (
-        app_conf.cloud.location if app_conf.cloud else DEFAULT_CODE_DIR
+        app_conf.cloud.code if app_conf.cloud else DEFAULT_CODE_DIR
     )
     template_code_file = code_dir / (app_conf.python_module + ".py.template")
     code_file = template_code_file.with_suffix("")
@@ -94,3 +102,119 @@ def scaffold(app_name: str, cloud_config: Optional[DefaultAppConfig] = None) -> 
     template_pytest_file.unlink()
 
     return conf_file
+
+
+def try_run_test_docker(
+    app_config: AppConf,
+    client,
+    docker_name,
+    container_name: str,
+    docker_config: TestDockerConfig,
+    test_path: Path,
+    secrets_path: Optional[Path],
+    sealed_secrets_path: Optional[Path],
+):
+    """Try to start the app docker to test"""
+    success = False
+    try:
+        container = run_app_test_docker(
+            client,
+            docker_name,
+            container_name,
+            docker_config,
+            app_config.healthcheck_endpoint,
+        )
+
+        success = run_tests(
+            app_config,
+            test_path,
+            secrets_path,
+            sealed_secrets_path,
+        )
+
+    except Exception as exc:
+        raise exc
+    finally:
+        try:
+            container = client.containers.get(container_name)
+            if not success:
+                LOG.info("The docker logs are:\n%s", container.logs().decode("utf-8"))
+            container.stop(timeout=1)
+            # We need to remove the container since we declare remove=False previously
+            container.remove()
+        except NotFound:
+            pass
+
+
+def run_app_test_docker(
+    client,
+    docker_name,
+    container_name: str,
+    docker_config: TestDockerConfig,
+    healthcheck_endpoint: str,
+) -> Container:
+    """Run the app docker to test."""
+    container = client.containers.run(
+        docker_name,
+        name=container_name,
+        command=docker_config.cmd(),
+        volumes=docker_config.volumes(),
+        entrypoint=TestDockerConfig.entrypoint,
+        ports=docker_config.ports(),
+        detach=True,
+        # We do not remove the container to be able to print the error (if some)
+        remove=False,
+    )
+
+    clock = ClockTick(
+        period=5,
+        timeout=10,
+        message="Test application docker is unreachable!",
+    )
+
+    while clock.tick():
+        # Note: container.status is not actualized.
+        # Get it again to have the current status
+        container = client.containers.get(container_name)
+
+        if container.status == "exited":
+            raise Exception("Application docker fails to start")
+
+        if is_ready(f"http://localhost:{docker_config.port}", healthcheck_endpoint):
+            break
+
+    return container
+
+
+def run_tests(
+    app_config: AppConf,
+    tests: Path,
+    secrets: Optional[Path],
+    sealed_secrets: Optional[Path],
+) -> bool:
+    """Run the tests."""
+
+    LOG.info("Installing tests requirements...")
+    for package in app_config.tests_requirements:
+        subprocess.check_call(
+            [sys.executable, "-m", "pip", "install", package],
+            stdout=subprocess.DEVNULL,
+        )
+
+    LOG.info("Running tests...")
+    env = dict(os.environ)
+    if secrets:
+        env["TEST_SECRET_JSON"] = str(secrets.resolve())
+
+    if sealed_secrets:
+        env["TEST_SEALED_SECRET_JSON"] = str(sealed_secrets.resolve())
+
+    try:
+        subprocess.check_call(app_config.tests_cmd, cwd=tests, env=env)
+
+        LOG.info("Tests successful")
+        return True
+    except subprocess.CalledProcessError:
+        LOG.error("Tests failed!")
+
+    return False
