+++ mse_cli/home/command/code_provider/localtest.py
@@ -1,17 +1,12 @@
-"""mse_cli.home.command.code_provider.test_dev module."""
+"""mse_cli.home.command.code_provider.localtest module."""
 
 import argparse
-import os
-import subprocess
-import sys
 from pathlib import Path
 from typing import Optional
 
-from docker.errors import BuildError, NotFound
-from docker.models.containers import Container
+from docker.errors import BuildError
 
-from mse_cli.core.bootstrap import is_ready
-from mse_cli.core.clock_tick import ClockTick
+from mse_cli.common.helpers import try_run_test_docker
 from mse_cli.core.conf import AppConf, AppConfParsingOption
 from mse_cli.core.test_docker import TestDockerConfig
 from mse_cli.home.command.helpers import get_client_docker
@@ -29,7 +24,7 @@ from mse_cli.log import LOGGER as LOG
 def add_subparser(subparsers):
     """Define the subcommand."""
     parser = subparsers.add_parser(
-        "test-dev", help="test a MSE app in a development context"
+        "localtest", help="test locally a MSE app in a development context"
     )
 
     parser.add_argument("--project", type=Path, help="path of the project to test")
@@ -140,7 +135,7 @@ def run(args) -> None:
         port=5000,
     )
 
-    try_run(
+    try_run_test_docker(
         code_config,
         client,
         docker_name,
@@ -150,122 +145,6 @@ def run(args) -> None:
         secrets_path,
         sealed_secrets_path,
     )
-
-
-def try_run(
-    app_config: AppConf,
-    client,
-    docker_name,
-    container_name: str,
-    docker_config: TestDockerConfig,
-    test_path: Path,
-    secrets_path: Optional[Path],
-    sealed_secrets_path: Optional[Path],
-):
-    """Try to start the app docker to test"""
-    success = False
-    try:
-        container = run_app_docker(
-            client,
-            docker_name,
-            container_name,
-            docker_config,
-            app_config.healthcheck_endpoint,
-        )
-
-        success = run_tests(
-            app_config,
-            test_path,
-            secrets_path,
-            sealed_secrets_path,
-        )
-
-    except Exception as exc:
-        raise exc
-    finally:
-        try:
-            container = client.containers.get(container_name)
-            if not success:
-                LOG.info("The docker logs are:\n%s", container.logs().decode("utf-8"))
-            container.stop(timeout=1)
-            # We need to remove the container since we declare remove=False previously
-            container.remove()
-        except NotFound:
-            pass
-
-
-def run_app_docker(
-    client,
-    docker_name,
-    container_name: str,
-    docker_config: TestDockerConfig,
-    healthcheck_endpoint: str,
-) -> Container:
-    """Run the app docker to test."""
-    container = client.containers.run(
-        docker_name,
-        name=container_name,
-        command=docker_config.cmd(),
-        volumes=docker_config.volumes(),
-        entrypoint=TestDockerConfig.entrypoint,
-        ports=docker_config.ports(),
-        detach=True,
-        # We do not remove the container to be able to print the error (if some)
-        remove=False,
-    )
-
-    clock = ClockTick(
-        period=5,
-        timeout=10,
-        message="Test application docker is unreachable!",
-    )
-
-    while clock.tick():
-        # Note: container.status is not actualized.
-        # Get it again to have the current status
-        container = client.containers.get(container_name)
-
-        if container.status == "exited":
-            raise Exception("Application docker fails to start")
-
-        if is_ready(f"http://localhost:{docker_config.port}", healthcheck_endpoint):
-            break
-
-    return container
-
-
-def run_tests(
-    app_config: AppConf,
-    tests: Path,
-    secrets: Optional[Path],
-    sealed_secrets: Optional[Path],
-) -> bool:
-    """Run the tests."""
-
-    LOG.info("Installing tests requirements...")
-    for package in app_config.tests_requirements:
-        subprocess.check_call(
-            [sys.executable, "-m", "pip", "install", package],
-            stdout=subprocess.DEVNULL,
-        )
-
-    LOG.info("Running tests...")
-    env = dict(os.environ)
-    if secrets:
-        env["TEST_SECRET_JSON"] = str(secrets.resolve())
-
-    if sealed_secrets:
-        env["TEST_SEALED_SECRET_JSON"] = str(sealed_secrets.resolve())
-
-    try:
-        subprocess.check_call(app_config.tests_cmd, cwd=tests, env=env)
-
-        LOG.info("Tests successful")
-        return True
-    except subprocess.CalledProcessError:
-        LOG.error("Tests failed!")
-
-    return False
 
 
 def build_test_docker(client, dockerfile: Path, docker_name: str):
